+++ tools/unencrypted-keys-gen/keygen/BUILD.bazel
@@ -1,0 +1,9 @@
+load("@prysm//tools/go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["keygen.go"],
+    importpath = "github.com/prysmaticlabs/prysm/tools/unencrypted-keys-gen/keygen",
+    visibility = ["//visibility:public"],
+    deps = ["@com_github_sirupsen_logrus//:go_default_library"],
+)
