+++ tools/unencrypted-keys-gen/main_test.go
@@ -3,6 +3,7 @@ package main
 import (
 	"bytes"
 	"encoding/json"
+	"github.com/prysmaticlabs/prysm/tools/unencrypted-keys-gen/keygen"
 	"reflect"
 	"testing"
 )
@@ -12,11 +13,11 @@ func TestSavesUnencryptedKeys(t *testing.T) {
 	numKeys = &keys
 	ctnr := generateUnencryptedKeys(0 /* start index */)
 	buf := new(bytes.Buffer)
-	if err := SaveUnencryptedKeysToFile(buf, ctnr); err != nil {
+	if err := keygen.SaveUnencryptedKeysToFile(buf, ctnr); err != nil {
 		t.Fatal(err)
 	}
 	enc := buf.Bytes()
-	dec := &UnencryptedKeysContainer{}
+	dec := &keygen.UnencryptedKeysContainer{}
 	if err := json.Unmarshal(enc, dec); err != nil {
 		t.Fatal(err)
 	}
